refactor: template 스태틱 메서드로 변경
