+++ docs/README.md
@@ -103,7 +103,7 @@
 - [x] 테스트 코드 매직 넘버 정리, `calledWith` 사용, 주석 표기
 - [x] 로깅 함수 삭제
 - [x] `RacingGame` 모델에서 불필요한 상태인 `movedInThisRound` 걷어내기
-- [ ] `template` 스태틱 메서드 형태로 변경해보기
+- [x] `template` 스태틱 메서드 형태로 변경해보기
 - [ ] 우승자 `alert`이 표시되기 전 다시 시작 시 `alert` 표시하지 않음
 
 ---
